Update el-segundo-scipion.xml
Changed the order of some stage directions. Problem: many in-line stage directions appear right of the line they refer to. As DraCor does not account for that in their layout, these stage directions have to be put in front of the line that they refer to. Otherwise, the HTML version of the play on the DraCor website is wrong.